[IMP] mail: move channel related code to discuss patch
This PR moves the channel-related code in thread_model.js to
discuss/thread_model_patch.js.

part of task-4675831
