+++ addons/mail/static/src/core/common/thread_model.js
@@ -76,24 +76,6 @@ export class Thread extends Record {
          */
         sort: (a1, a2) => (a1.id < a2.id ? 1 : -1),
     });
-    get allowedToLeaveChannelTypes() {
-        return ["channel", "group"];
-    }
-    get canLeave() {
-        return (
-            this.allowedToLeaveChannelTypes.includes(this.channel_type) &&
-            this.group_ids.length === 0 &&
-            this.store.self?.type === "partner"
-        );
-    }
-    get allowedToUnpinChannelTypes() {
-        return ["chat"];
-    }
-    get canUnpin() {
-        return (
-            this.parent_channel_id || this.allowedToUnpinChannelTypes.includes(this.channel_type)
-        );
-    }
     /** @type {boolean} */
     can_react = true;
     chat_window = fields.One("ChatWindow", {
@@ -119,19 +101,6 @@ export class Thread extends Record {
     description;
     /** @type {string} */
     display_name;
-    displayToSelf = fields.Attr(false, {
-        compute() {
-            return (
-                this.is_pinned ||
-                (["channel", "group"].includes(this.channel_type) &&
-                    this.hasSelfAsMember &&
-                    !this.parent_channel_id)
-            );
-        },
-        onUpdate() {
-            this.onPinStateUpdated();
-        },
-    });
     followers = fields.Many("mail.followers", {
         /** @this {import("models").Thread} */
         onAdd(r) {
@@ -323,18 +292,6 @@ export class Thread extends Record {
         return this.needactionMessages.length > 0;
     }
 
-    get typesAllowingCalls() {
-        return ["chat", "channel", "group"];
-    }
-
-    get allowCalls() {
-        return (
-            !this.isTransient &&
-            this.typesAllowingCalls.includes(this.channel_type) &&
-            !this.correspondent?.persona.eq(this.store.odoobot)
-        );
-    }
-
     /**
      * Return the name of the given persona to display in the context of this
      * thread.
@@ -345,17 +302,10 @@ export class Thread extends Record {
         return persona.displayName;
     }
 
-    get hasAttachmentPanel() {
-        return this.model === "discuss.channel";
-    }
-
-    get isChatChannel() {
-        return ["chat", "group"].includes(this.channel_type);
-    }
-
     get supportsCustomChannelName() {
         return this.isChatChannel && this.channel_type !== "group";
     }
+
     get displayName() {
         return this.display_name;
     }
@@ -366,10 +316,6 @@ export class Thread extends Record {
 
     get avatarUrl() {
         return this.module_icon ?? this.store.DEFAULT_AVATAR;
-    }
-
-    get allowDescription() {
-        return ["channel", "group"].includes(this.channel_type);
     }
 
     get isTransient() {
@@ -423,13 +369,6 @@ export class Thread extends Record {
 
     onPinStateUpdated() {}
 
-    get invitationLink() {
-        if (!this.uuid || this.channel_type === "chat") {
-            return undefined;
-        }
-        return `${window.location.origin}/chat/${this.id}/${this.uuid}`;
-    }
-
     get isEmpty() {
         return this.messages.length === 0;
     }
@@ -448,15 +387,6 @@ export class Thread extends Record {
 
     get rpcParams() {
         return {};
-    }
-
-    executeCommand(command, body = "") {
-        return this.store.env.services.orm.call(
-            "discuss.channel",
-            command.methodName,
-            [[this.id]],
-            { body }
-        );
     }
 
     /** @param {{after: Number, before: Number}} */
@@ -640,12 +570,6 @@ export class Thread extends Record {
         return "/mail/thread/messages";
     }
 
-    async leave() {
-        await this.store.env.services.orm.silent.call("discuss.channel", "action_unfollow", [
-            this.id,
-        ]);
-    }
-
     /**
      * Get ready to jump to a message in a thread. This method will fetch the
      * messages around the message to jump to if required, and update the thread
@@ -698,12 +622,6 @@ export class Thread extends Record {
         this.message_needaction_counter = 0;
     }
 
-    async markAsFetched() {
-        await this.store.env.services.orm.silent.call("discuss.channel", "channel_fetched", [
-            [this.id],
-        ]);
-    }
-
     /**
      * @param {Object} [options] used in overrides
      */
@@ -718,24 +636,6 @@ export class Thread extends Record {
         if (this.message_needaction_counter > 0) {
             this.markAllMessagesAsRead();
         }
-    }
-
-    /** @param {string} data base64 representation of the binary */
-    async notifyAvatarToServer(data) {
-        await rpc("/discuss/channel/update_avatar", {
-            channel_id: this.id,
-            data,
-        });
-    }
-
-    async notifyDescriptionToServer(description) {
-        this.description = description;
-        return this.store.env.services.orm.call(
-            "discuss.channel",
-            "channel_change_description",
-            [[this.id]],
-            { description }
-        );
     }
 
     /** @param {import("models").Message} message */
@@ -761,48 +661,6 @@ export class Thread extends Record {
         await this.store.chatHub.initPromise;
         const chatWindow = this.store.ChatWindow.get({ thread: this });
         await chatWindow?.close({ notifyState: false, ...options });
-    }
-
-    pin() {
-        if (this.model !== "discuss.channel" || this.store.self.type !== "partner") {
-            return;
-        }
-        this.is_pinned = true;
-        return this.store.env.services.orm.silent.call(
-            "discuss.channel",
-            "channel_pin",
-            [this.id],
-            { pinned: true }
-        );
-    }
-
-    /** @param {string} name */
-    async rename(name) {
-        const newName = name.trim();
-        if (
-            newName !== this.displayName &&
-            ((newName && this.channel_type === "channel") || this.isChatChannel)
-        ) {
-            if (this.channel_type === "channel" || this.channel_type === "group") {
-                this.name = newName;
-                await this.store.env.services.orm.call(
-                    "discuss.channel",
-                    "channel_rename",
-                    [[this.id]],
-                    { name: newName }
-                );
-            } else if (this.supportsCustomChannelName) {
-                if (this.selfMember) {
-                    this.selfMember.custom_channel_name = newName;
-                }
-                await this.store.env.services.orm.call(
-                    "discuss.channel",
-                    "channel_set_custom_name",
-                    [[this.id]],
-                    { name: newName }
-                );
-            }
-        }
     }
 
     addOrReplaceMessage(message, tmpMsg) {
@@ -909,26 +767,6 @@ export class Thread extends Record {
                 this.messages.splice(afterIndex - 1, 0, message);
             }
         }
-    }
-
-    async leaveChannel({ force = false } = {}) {
-        if (
-            this.channel_type !== "group" &&
-            this.create_uid?.eq(this.store.self.main_user_id) &&
-            !force
-        ) {
-            await this.askLeaveConfirmation(
-                _t("You are the administrator of this channel. Are you sure you want to leave?")
-            );
-        }
-        if (this.channel_type === "group" && !force) {
-            await this.askLeaveConfirmation(
-                _t(
-                    "You are about to leave this group conversation and will no longer have access to it unless you are invited again. Are you sure you want to continue?"
-                )
-            );
-        }
-        this.leave();
     }
 
     _getActualModelName() {
